refactor(core): remove UniversalOpsSourceMacros, inline sourceCode/withSourceCode impls in UniversalOps companion and update comments/docs
